updated odds

+++ src/autobet.py
@@ -66,7 +66,7 @@ class Autobet:
     def place_bet(self, best_slips, profile_id, token):
         balance = self.get_bal(token)
         stakeable = balance/2
-        stake = int(stakeable/len(best_slips))
+        stake = stakeable/len(best_slips)
         for bet_slip in best_slips:
             body = '{'
             body = body + f'''
